Move  to testing/. (#1067)
